requirements.txt
# Web framework
fastapi>=0.97.0
uvicorn[standard]>=0.22.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-multipart>=0.0.6
pydantic>=1.10.8
dynaconf>=3.2.0
loguru>=0.7.0
python-dateutil>=2.8.2
httpx>=0.24.1
aiofiles>=23.1.0

# OCR and document processing
pytesseract>=0.3.10
pdf2image>=1.16.3
pdfplumber>=0.9.0
Pillow>=9.5.0

# Machine learning
scikit-learn>=1.2.2
PyYAML>=6.0

# Monitoring
psutil>=5.9.5

# For development/testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0